basic crew funcitons tested and working

+++ src/mutations/recipeBook.js
@@ -14,11 +14,11 @@ import {
   createPermissionOnRecipeBook,
   deleteRecipeBookPermission
 } from "../actions/recipeBook.js";
+import { resolvePermission } from "../actions/utils.js";
 
 async function newRecipeBook(parent, args, context, info) {
   const { userId } = context;
-
-  const recipeBook = await createRecipeBook(
+  const { status, permission, recipeBook } = await createRecipeBook(
     context,
     args.name,
     args.description,
@@ -26,14 +26,22 @@ async function newRecipeBook(parent, args, context, info) {
   );
 
   return {
-    status: recipeBook.status,
-    permission: recipeBook.permission,
-    recipeBook: recipeBook.recipeBook
+    status,
+    permission,
+    recipeBook
   };
 }
 
 async function editRecipeBook(parent, args, context, info) {
-  const recipeBook = await updateRecipeBook(
+  if (!resolvePermission(args.permission, "Manager")) {
+    return {
+      status: {
+        code: "Failure",
+        message: ""
+      }
+    };
+  }
+  const { status, recipeBook } = await updateRecipeBook(
     context,
     args.recipeBookId,
     args.name,
@@ -41,23 +49,38 @@ async function editRecipeBook(parent, args, context, info) {
     args.permission
   );
 
-  return {
-    status: recipeBook.status,
-    permission: recipeBook.permission,
-    recipeBook: recipeBook.recipeBook
-  };
+  return { status, permission: args.permission, recipeBook };
 }
 
 async function trashRecipeBook(parent, args, context, info) {
-  const statusMessage = await deleteRecipeBook(
+  if (!resolvePermission(permission, "Owner")) {
+    return {
+      status: {
+        code: "Failure",
+        message: ""
+      }
+    };
+  }
+  const { recipeBook, status } = await deleteRecipeBook(
     context,
     args.recipeBookId,
     args.permission
   );
-  return statusMessage;
+  return { recipeBook, status };
 }
 
 async function addBuildToRecipeBook(parent, args, context, info) {
+  if (
+    !resolvePermission(args.bookPermission, "Manager") ||
+    !resolvedPermission(args.buildPermission, "Manager")
+  ) {
+    return {
+      status: {
+        message: "You don't have permission to add to this Recipe Book",
+        code: "Failure"
+      }
+    };
+  }
   return await createBuildOnRecipeBook(
     context,
     args.buildId,
@@ -68,6 +91,15 @@ async function addBuildToRecipeBook(parent, args, context, info) {
 }
 
 async function removeBuildFromRecipeBook(parent, args, context, info) {
+  if (!resolvePermission(permission, "Manager")) {
+    return {
+      status: {
+        message:
+          "You don't have permission to remove anything from this Recipe Book",
+        code: "Failure"
+      }
+    };
+  }
   return await deleteBuildFromRecipeBook(
     context,
     args.buildId,
@@ -77,6 +109,15 @@ async function removeBuildFromRecipeBook(parent, args, context, info) {
 }
 
 async function changeRecipeBookPermission(parent, args, context, info) {
+  if (!resolvePermission(userPermission, permission)) {
+    return {
+      status: {
+        message:
+          "You don't have permission to remove anything from this Recipe Book",
+        code: "Failure"
+      }
+    };
+  }
   const recipeBook = await createPermissionOnRecipeBook(
     context,
     args.userId,
@@ -85,13 +126,16 @@ async function changeRecipeBookPermission(parent, args, context, info) {
     args.userPermission
   );
   console.log(recipeBook);
-  return {
-    status: { message: "barf", code: "success" },
-    recipeBookId: recipeBook.recipeBookId
-  };
+  return recipeBook;
 }
 
 async function removeRecipeBookPermission(parent, args, context, info) {
+  if (!resolvePermission(permission, "Manager")) {
+    return {
+      message: "You don't have permission to remove users from this book!",
+      code: "Failure"
+    };
+  }
   const response = await deleteRecipeBookPermission(
     context,
     args.userId,
